add a CMakeLists.txt system for building projects
This is following the recommanded way by RaspberryPi,
with the difference that we replace the fetch script
by our own, always fetching the pico-sdk and pico-ice-sdk.

The import script is very short, it would be trivial to tune
it to point at an external directory instead.

The user would write its own CMakeLists.txt, which would
include(cmake/pico_ice_sdk_init.cmake), itself configuring
as much as possible of the pico-sdk.