fix custom board definition
This took more time than it looks, as it is not documented
that it is required to setup PICO_CONFIG_HEADER_FILES for
this to work, and there is no error message when not set.

It is also not working to just define the .h include dir,
the pico-sdk really wants the .cmake include dir set as
well.