simplify the flash API and fix bit-banged SPI
The flash API now has hardcoded paths: we only have one flash chip
onboard, and there might be a new smem_*() API coming as a back-end.

The timing of SPI was off by half a clock.

A write operation always seems to work, but an erase seems to get stuck
at a single page in a row before a repeated erase. That still needs to
be troubleshooted.