fix USB UART forwarding again
the pico-sdk is also implementing tud_cdc_rx_cb(), which overrides our own
so we need to disable pico_stdio_usb as well in addition to disable
pico_stdio_uart.

It should work nicely since we do not need interrupts on stdin and stdout
works without any interrupt or callback.

It seems like we are lucky and that the pico-sdk allows us to do everything
we need to.