keep DMA interrupt enabled during DFU flash programming
Since commit 2b52c6e (Use hardware SPI and DMA transfer), hardware SPI
has been used with hardware DMA to program the ICE40 FPGA SPI flash.
This means that in order for SPI transfers to the flash to complete,
the DMA interrupt must remain enabled. Keep the DMA interrupt enabled
when entering the USB DFU flash programming code, to ensure that SPI
flash writes complete.

Without this fix, dfu-util is unable to program the FLASH slot, and can
only program to the CRAM slot. This is because the SPI transfer will
wait indefinitely to complete (as DMA interrupt is disabled), and the
watchdog will reset the RT2040.

Signed-off-by: Daniel DeGrasse <daniel@degrasse.com>