[draft] uart1 fifo
this implements rudimentary fifo for uart1 only
tud_cdc_n_write_char aquires mutex and it's not recommended in ISRs (can deadlock for example, is long)
fifo has two sets of variables, index and total, read and written. This is to limit race conditions.
Each set is accessed only in one context.

Fixes #53