cmake: add a warning so that users of new board do not forget to adjust
By default, the cmake helpers select the pico_ice board. This board
is still widespread, but not manufactured anymore.

New users will get started with the pico2-ice, try to find a compromise
for current users and new users: default does not change, new users get
a warning so they remember to set the -D flag.